refactor: blg review: rename internal/bufio.PacketReader into internal/buffr.BufferedReader
