+++ internal/buffr/buffr.go
@@ -1,5 +1,5 @@
-// Package bufio provides specialised packet buffers.
-package bufio
+// Package buffr provides a buffered reader.
+package buffr
 
 import (
 	"bytes"
@@ -10,18 +10,18 @@ import (
 // ErrNegativeCount is returned when a negative number of bytes is requested.
 var ErrNegativeCount = errors.New("bufio: negative count")
 
-// PacketReader implements buffering for an io.Reader object.
-// PacketReader will not read from the underlying reader unless its
+// BufferedReader implements buffering for an io.Reader object.
+// BufferedReader will not read from the underlying reader unless its
 // buffer is empty, reducing the risk of cross packet reading.
-type PacketReader struct {
+type BufferedReader struct {
 	rd   io.Reader
 	buf  []byte
 	r, w int
 }
 
-// NewPacketReader returns a new PacketReader whose buffer is size.
-func NewPacketReader(rd io.Reader, size int) *PacketReader {
-	r := &PacketReader{}
+// NewBufferedReader returns a new BufferedReader whose buffer is size.
+func NewBufferedReader(rd io.Reader, size int) *BufferedReader {
+	r := &BufferedReader{}
 	r.reset(make([]byte, size), rd)
 	return r
 }
@@ -29,15 +29,15 @@ func NewPacketReader(rd io.Reader, size int) *PacketReader {
 // Reset discards any buffered data, resets all state, and switches
 // the buffered reader to read from rd.
 // Calling Reset on the zero value of Reader will panic.
-func (r *PacketReader) Reset(rd io.Reader) {
+func (r *BufferedReader) Reset(rd io.Reader) {
 	if r.buf == nil {
 		panic(errors.New("bufio: uninitialized PacketBuffer"))
 	}
 	r.reset(r.buf, rd)
 }
 
-func (r *PacketReader) reset(buf []byte, rd io.Reader) {
-	*r = PacketReader{
+func (r *BufferedReader) reset(buf []byte, rd io.Reader) {
+	*r = BufferedReader{
 		buf: buf,
 		rd:  rd,
 	}
@@ -45,7 +45,7 @@ func (r *PacketReader) reset(buf []byte, rd io.Reader) {
 
 const maxConsecutiveEmptyReads = 3
 
-func (r *PacketReader) fill() error {
+func (r *BufferedReader) fill() error {
 	// Ignore any data in the buffer.
 	r.r = 0
 	r.w = 0
@@ -64,13 +64,13 @@ func (r *PacketReader) fill() error {
 }
 
 // Buffered returns the number of bytes that can be read from the current buffer.
-func (r *PacketReader) Buffered() int {
+func (r *BufferedReader) Buffered() int {
 	return r.w - r.r
 }
 
 // Peek reads data in p without advancing the reader.
 // It returns the number of bytes read into p.
-func (r *PacketReader) Peek(p []byte) (int, error) {
+func (r *BufferedReader) Peek(p []byte) (int, error) {
 	if r.r == r.w {
 		err := r.fill()
 		if err != nil {
@@ -84,7 +84,7 @@ func (r *PacketReader) Peek(p []byte) (int, error) {
 
 // Discard skips the next n bytes, returning the number of bytes discarded.
 // If Discard skips fewer than n bytes, it also returns an error.
-func (r *PacketReader) Discard(n int) (int, error) {
+func (r *BufferedReader) Discard(n int) (int, error) {
 	if n < 0 {
 		return 0, ErrNegativeCount
 	}
@@ -104,7 +104,7 @@ func (r *PacketReader) Discard(n int) (int, error) {
 
 // Read reads data in p.
 // It returns the number of bytes read into p.
-func (r *PacketReader) Read(p []byte) (int, error) {
+func (r *BufferedReader) Read(p []byte) (int, error) {
 	if r.r == r.w {
 		err := r.fill()
 		if err != nil {
@@ -118,7 +118,7 @@ func (r *PacketReader) Read(p []byte) (int, error) {
 }
 
 // ReadByte reads the next byte.
-func (r *PacketReader) ReadByte() (byte, error) {
+func (r *BufferedReader) ReadByte() (byte, error) {
 	if r.r == r.w {
 		err := r.fill()
 		if err != nil {
@@ -135,7 +135,7 @@ func (r *PacketReader) ReadByte() (byte, error) {
 // returning a slice pointing at the bytes in the buffer.
 // If ReadSlice encounters the end of the buffer before finding the delimiter,
 // it returns an io.EOF error.
-func (r *PacketReader) ReadSlice(delim byte) ([]byte, error) {
+func (r *BufferedReader) ReadSlice(delim byte) ([]byte, error) {
 	if r.r == r.w {
 		err := r.fill()
 		if err != nil {
@@ -152,54 +152,4 @@ func (r *PacketReader) ReadSlice(delim byte) ([]byte, error) {
 	line := r.buf[r.r : r.r+i+1]
 	r.r += i + 1
 	return line, nil
-}
-
-// PacketWriter implements buffering for an io.Writer object.
-// PacketWriter will grow to the needed size and will not
-// writer to the io.Writer until Flush is called.
-type PacketWriter struct {
-	wr  io.Writer
-	buf []byte
-}
-
-// NewPacketWriter returns a new PacketWriter.
-func NewPacketWriter(wr io.Writer) *PacketWriter {
-	return &PacketWriter{wr: wr}
-}
-
-// Reset discards any buffered data, resets all state, and switches
-// the buffered writer to write to wr.
-func (w *PacketWriter) Reset(wr io.Writer) {
-	w.wr = wr
-	if w.buf != nil {
-		w.buf = w.buf[:0]
-	}
-}
-
-// Write writes p to the buffer.
-func (w *PacketWriter) Write(p []byte) (int, error) {
-	w.buf = append(w.buf, p...)
-	return len(p), nil
-}
-
-// WriteByte writes the byte b to the buffer.
-func (w *PacketWriter) WriteByte(b byte) error {
-	w.buf = append(w.buf, b)
-	return nil
-}
-
-// Flush writes the buffer to the writer and resets all state.
-func (w *PacketWriter) Flush() error {
-	if len(w.buf) == 0 {
-		return nil
-	}
-	n, err := w.wr.Write(w.buf)
-	if n < len(w.buf) && err == nil {
-		err = io.ErrShortWrite
-	}
-	if err != nil {
-		return err
-	}
-	w.buf = w.buf[:0]
-	return nil
 }
